Design: 게시판 카테고리 삭제 버튼 아이콘 -> 텍스트 변경
